Categories List endpoint

+++ Application/Infrastructure/Entity/Transaction/Transaction.cs
@@ -1,8 +1,8 @@
 ﻿namespace MMX.Infrastructure.Entity.Transaction
 {
     using MMX.Common.Domain;
+    using MMX.Domain.Enum;
     using MMX.Infrastructure.Entity.Category;
-    using MMX.Infrastructure.Enum;
 
     public class Transaction : BaseCreatedEntity
     {
@@ -14,7 +14,7 @@
 
         public string? Note { get; set; }
 
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public virtual Category? Category { get; set; }
     }
